+++ include/iot/logger.h
@@ -22,63 +22,68 @@ extern "C" {
 #define IOT_LOGGER_TYPE "IOT::Logger"
 /** Default log level warning */
 #define IOT_LOGLEVEL_DEFAULT IOT_LOG_WARN
-
-/** Logger structure */
-struct iot_logger_t;
+/** Maximum log message size */
+#define IOT_LOG_MSG_MAX 1024
 
 /**
- * Alias for log level enumeration
+ * Log level enumeration
  */
 typedef enum iot_loglevel_t
 {
-  IOT_LOG_NONE = 0,   /**< No logger */
-  IOT_LOG_ERROR,      /**< Log level 1 */
-  IOT_LOG_WARN,       /**< Log level 2 */
-  IOT_LOG_INFO,       /**< Log level 3 */
-  IOT_LOG_DEBUG,      /**< Log level 4 */
-  IOT_LOG_TRACE       /**< Log level 5 - max IOT debug level */
+  IOT_LOG_NONE = 0,   /**< No logging */
+  IOT_LOG_ERROR,      /**< Error logging */
+  IOT_LOG_WARN,       /**< Warning and Error logging */
+  IOT_LOG_INFO,       /**< Information, Warning and Error logging */
+  IOT_LOG_DEBUG,      /**< Debug, Information, Warning and Error logging */
+  IOT_LOG_TRACE       /**< Trace, Debug, Information, Warning and Error logging */
 } iot_loglevel_t;
 
-/** Alias for IOT logger function pointer */
-typedef void (*iot_log_function_t) (struct iot_logger_t * logger, iot_loglevel_t level, uint64_t timestamp, const char * message);
-
-/* Default set of supported logger implementation functions */
-
-#ifdef IOT_HAS_FILE
 /**
- * @brief Log message to a file
- *
- * @param logger     Pointer to a logger component, to get the component name. "default" is used if logger is NULL
- * @param level      Log level
- * @param timestamp  Timestamp information to log
- * @param message    Log message
+ * Log type enumeration
  */
-extern void iot_log_file (struct iot_logger_t * logger, iot_loglevel_t level, uint64_t timestamp, const char * message);
-#endif
+typedef enum iot_logger_type_t
+{
+  IOT_LOGGER_CONSOLE, /**< Console logger to stdout or stderr (for Warning and Error) */
+  IOT_LOGGER_FILE,    /**< File logger */
+  IOT_LOGGER_UDP,     /**< UDP logger (to fixed address or broadcast) */
+  IOT_LOGGER_CUSTOM   /**< Non core, custom logger */
+} iot_logger_type_t;
 
 /**
- * @brief Log message to console
- *
- * @param logger     Pointer to a logger component, to get the component name. "default" is used if logger is NULL
- * @param level      Log level
- * @param timestamp  Timestamp information to log
- * @param message    Log message
- */
-extern void iot_log_console (struct iot_logger_t * logger, iot_loglevel_t level, uint64_t timestamp, const char * message);
-
-/**
- * Alias for logger structure
+ * Public logger struct. Do not use directly or stack allocate.
  */
 typedef struct iot_logger_t
 {
-  iot_component_t component;      /**< Details of a given logger component */
+  iot_component_t component;      /**< Component base */
   volatile iot_loglevel_t level;  /**< Log level */
-  volatile iot_loglevel_t save;   /**< Last saved log level */
-  char * name;                    /**< Name of logger */
-  char * to;                      /**< Log 'to' a certain file */
-  iot_log_function_t impl;        /**< Pointer to function that handles logging functionality */
-  struct iot_logger_t * next;     /**< Pointer to next logger structure */
 } iot_logger_t;
+
+/**
+ * Logger implementation function type
+ */
+typedef void (*iot_log_function_t) (struct iot_logger_t * logger, iot_loglevel_t level, uint64_t timestamp, const char * message);
+
+
+#if defined (IOT_HAS_FILE) && !defined (_AZURESPHERE_)
+/** File logger function */
+extern void iot_log_file (struct iot_logger_t * logger, iot_loglevel_t level, uint64_t timestamp, const char * message);
+#endif
+/** Console logger function */
+extern void iot_log_console (struct iot_logger_t * logger, iot_loglevel_t level, uint64_t timestamp, const char * message);
+/** UDP logger function */
+extern void iot_log_udp (struct iot_logger_t * logger, iot_loglevel_t level, uint64_t timestamp, const char * message);
+
+/**
+ * @brief Allocate memory an initialize logger component
+ *
+ * The function to allocate memory and initialize logger component. The logger component logs messages to console
+ *
+ * @param name        Identifier to use for logging
+ * @param level       Log level
+ * @param self_start  'true' will start the logger after initialization
+ * @return            Pointer to the logger component created
+ */
+extern iot_logger_t * iot_logger_alloc (const char * name, iot_loglevel_t level, bool self_start);
 
 /**
  * @brief Allocate memory and initialize logger component with the custom log function
@@ -94,18 +99,6 @@ typedef struct iot_logger_t
  * @return            Pointer to the logger component created
  */
 extern iot_logger_t * iot_logger_alloc_custom (const char * name, iot_loglevel_t level, const char * to, iot_log_function_t impl, iot_logger_t * next, bool self_start);
-
-/**
- * @brief Allocate memory an initialize logger component
- *
- * The function to allocate memory and initialize logger component. The logger component logs messages to console
- *
- * @param name        Identifier to use for logging
- * @param level       Log level
- * @param self_start  'true' will start the logger after initialization
- * @return            Pointer to the logger component created
- */
-extern iot_logger_t * iot_logger_alloc (const char * name, iot_loglevel_t level, bool self_start);
 
 /**
  * @brief Increment the logger reference count
@@ -142,14 +135,6 @@ extern void iot_logger_stop (iot_logger_t * logger);
  * @return  Pointer to the logging component
  */
 extern iot_logger_t * iot_logger_default (void);
-
-/**
- * @brief Get next logger
- *
- * @param logger  Pointer to the logger component
- * @return        Pointer to the next logger component stored by the logger
- */
-extern iot_logger_t * iot_logger_next (iot_logger_t * logger);
 
 /* Logging functions */
 /**