+++ scripts/package.sh
@@ -1,8 +1,7 @@
 #!/bin/sh
+set -e -x
 
 # Build APK, DEB or RPM packages
-
-set -e -x
 
 # Process arguments
 
@@ -25,11 +24,6 @@ do
   esac
 done
 
-if [ -f /etc/os-release ]
-then
-  SYSTEM=$(sed -e 's/^ID=\(.*\)/\1/;t;d' < /etc/os-release | tr -d \")
-fi
-
 BROOT="${ROOT}/${BARCH}"
 VER=$(cat ${ROOT}/VERSION)
 PKG_VER=$(cut -d . -f 1,2 < ${ROOT}/VERSION)
@@ -39,11 +33,6 @@ DESC_MAIN="IOT C Framework"
 DESC_DEV="IOT C Framework development"
 DESC_DBG="IOT C Framework (debug enabled)"
 FPM=fpm
-
-if [ "${SYSTEM}" = "opensuse-leap" ]
-then
-  FPM=fpm.ruby2.5
-fi
 
 build_apk()
 {
@@ -61,7 +50,7 @@ build_apk()
 }
 
 case ${SYSTEM} in
-  alpine)
+  alpine*)
     cd ${ROOT}
     case ${BARCH} in
       arm64)
@@ -89,7 +78,7 @@ case ${SYSTEM} in
     export DEPS=libuuid
     build_apk "${BROOT}/debug" "iotech-iot-dev-${PKG_VER}-${VER}_${OS_ARCH}"
     ;;
-  debian|ubuntu)
+  debian*|ubuntu*)
     OS_ARCH=$(dpkg --print-architecture)
     cd ${ROOT}/${BARCH}/release
 
@@ -126,13 +115,13 @@ case ${SYSTEM} in
 
     rm *.tar.gz
     ;;
-  photon|centos|fedora|opensuse*)
+  photon*|centos*|fedora*|opensuse*)
     case ${BARCH} in
       arm64)
         OS_ARCH=aarch64
         ;;
       arm32)
-        if [ "${SYSTEM}" = "opensuse-leap" ]
+        if [ "${SYSTEM}" = "opensuse-15.3" ]
         then
           OS_ARCH=armv7hl
         else
@@ -143,28 +132,46 @@ case ${SYSTEM} in
         OS_ARCH=${BARCH}
         ;;
     esac
-    if [ "${SYSTEM}" = "opensuse-leap" ]
-    then
-      UUID_LIB=libuuid1
-    else
-      UUID_LIB=libuuid
-    fi
+
+    case ${SYSTEM} in
+      photon-40)
+        RPM_DIST=ph4
+        UUID_DEV_DEP=util-linux-devel
+      ;;
+      fedora-34)
+        RPM_DIST=fc34
+        UUID_DEP=libuuid1
+        UUID_DEV_DEP=libuuid-devel
+      ;;
+      centos-8)
+        RPM_DIST=el8
+        UUID_DEP=libuuid
+        UUID_DEV_DEP=libuuid-devel
+      ;;
+      opensuse-15.3)
+        FPM=fpm.ruby2.5
+        UUID_DEP=libuuid1
+        UUID_DEV_DEP=libuuid-devel
+      ;;
+    esac
+
     cd ${ROOT}/${BARCH}/release
 
     ${FPM} -s dir -t rpm -n iotech-iot-${PKG_VER} -v "${VER}" \
       -C _CPack_Packages/Linux/TGZ/iotech-iot-${PKG_VER}-${VER}_${OS_ARCH} \
+      --architecture "${OS_ARCH}" ${RPM_DIST:+--rpm-dist ${RPM_DIST}} \
       --prefix /opt/iotech/iot \
       --description "${DESC_MAIN}" \
       --vendor "IOTech" --maintainer "${MAINT_EMAIL}" \
-      --exclude include --exclude docs --exclude examples --exclude *.a \
-      --depends ${UUID_LIB}
+      --exclude include --exclude docs --exclude examples --exclude *.a ${UUID_DEP:+--depends ${UUID_DEP}}
 
     ${FPM} -s dir -t rpm -n iotech-iot-${PKG_VER}-dev -v "${VER}" \
       -C _CPack_Packages/Linux/TGZ/iotech-iot-${PKG_VER}-${VER}_${OS_ARCH} \
+      --architecture "${OS_ARCH}" ${RPM_DIST:+--rpm-dist ${RPM_DIST}} \
       --prefix /opt/iotech/iot \
       --description "${DESC_DEV}" \
       --vendor "IOTech" --maintainer "${MAINT_EMAIL}" \
-      --exclude lib \
+      --exclude lib ${UUID_DEV_DEP:+--depends ${UUID_DEV_DEP}} \
       --depends iotech-iot-${PKG_VER}
 
     rm *.tar.gz
@@ -173,11 +180,11 @@ case ${SYSTEM} in
 
     ${FPM} -s dir -t rpm -n iotech-iot-${PKG_VER}-dbg -v "${VER}" \
       -C _CPack_Packages/Linux/TGZ/iotech-iot-dev-${PKG_VER}-${VER}_${OS_ARCH} \
+      --architecture "${OS_ARCH}" ${RPM_DIST:+--rpm-dist ${RPM_DIST}} \
       --prefix /opt/iotech/iot \
       --description "${DESC_DBG}" \
       --vendor "IOTech" --maintainer "${MAINT_EMAIL}" \
-      --depends ${UUID_LIB} \
-      --conflicts iotech-iot-${PKG_VER} --conflicts iotech-iot-${PKG_VER}-dev
+      --conflicts iotech-iot-${PKG_VER} --conflicts iotech-iot-${PKG_VER}-dev ${UUID_DEV_DEP:+--depends ${UUID_DEV_DEP}}
 
     rm *.tar.gz
     ;;