+++ src/c/iot.c
@@ -120,6 +120,12 @@ bool iot_file_write_binary (const char * path, const uint8_t * binary, size_t le
   return ok;
 }
 
+extern bool iot_file_delete (const char * path)
+{
+  (void) path;
+  return (Storage_DeleteMutableFile () == 0);
+}
+
 #else
 
 extern bool iot_file_delete (const char * path)