+++ src/c/logger.c
@@ -1,5 +1,5 @@
 //
-// Copyright (c) 2018,2019 IOTech
+// Copyright (c) 2018-2021 IOTech
 //
 // SPDX-License-Identifier: Apache-2.0
 //
@@ -8,6 +8,9 @@
 #include "iot/container.h"
 #include "iot/time.h"
 #include <stdarg.h>
+#include <sys/socket.h>
+#include <netinet/in.h>
+#include <arpa/inet.h>
 
 #ifdef IOT_HAS_PRCTL
 #include <sys/prctl.h>
@@ -32,33 +35,49 @@ static time_t time (time_t *t)
 #define IOT_LOGGER_FACTORY NULL
 #endif
 
+typedef struct iot_logger_impl_t
+{
+  iot_logger_t base;                  // Public part of logger
+  volatile iot_loglevel_t save;       // Last saved log level
+  char * name;                        // Name of logger
+  iot_log_function_t impl;            // Log implementation function
+  struct iot_logger_impl_t * next;    // Pointer to next logger (can be chained in config)
+  int sock;                           // Socket for UDP logger
+  struct sockaddr_in addr;            // Address for UDP logger
+  char buff [IOT_LOG_MSG_MAX];        // Log format buffer
+#if defined (IOT_HAS_FILE) && !defined (_AZURESPHERE_)
+  FILE * fd;                          // Descriptor for File logger
+#endif
+}
+iot_logger_impl_t;
+
 static const char * iot_log_levels[IOT_LOG_LEVELS] = {"", "ERROR", "WARN", "Info", "Debug", "Trace"};
-static iot_logger_t iot_logger_dfl;
+static iot_logger_impl_t iot_logger_dfl;
 
 iot_logger_t * iot_logger_default (void)
 {
   static iot_logger_t * logger = NULL;
   if (logger == NULL)
   {
-    logger = &iot_logger_dfl;
+    logger = &iot_logger_dfl.base;
     memset (&iot_logger_dfl, 0, sizeof (iot_logger_dfl));
     iot_component_init (&logger->component, IOT_LOGGER_FACTORY, (iot_component_start_fn_t) iot_logger_start, (iot_component_stop_fn_t) iot_logger_stop);
-    iot_logger_dfl.level = iot_logger_dfl.save = IOT_LOGLEVEL_DEFAULT;
+    iot_logger_dfl.base.level = iot_logger_dfl.save = IOT_LOGLEVEL_DEFAULT;
     iot_logger_dfl.impl = iot_log_console;
   }
   return logger;
 }
 
-static void iot_logger_log (iot_logger_t *logger, iot_loglevel_t level, va_list args)
+static void iot_logger_log (iot_logger_impl_t * logger, iot_loglevel_t level, va_list args)
 {
   char str[1024];
   const char * fmt = va_arg (args, const char *);
   vsnprintf (str, sizeof (str), fmt, args);
   uint64_t ts = iot_time_usecs ();
-  if (logger->level >= level) (logger->impl) (logger, level, ts, str);
+  if (logger->base.level >= level) (logger->impl) (&logger->base, level, ts, str);
   while ((logger = logger->next))
   {
-    if (logger->level >= level) (logger->impl) (logger, level, ts, str);
+    if (logger->base.level >= level) (logger->impl) (&logger->base, level, ts, str);
   }
 }
 
@@ -66,7 +85,7 @@ void iot_log__error (iot_logger_t * logger, ...)
 {
   va_list args;
   va_start (args, logger);
-  iot_logger_log (logger, IOT_LOG_ERROR, args);
+  iot_logger_log ((iot_logger_impl_t*) logger, IOT_LOG_ERROR, args);
   va_end (args);
 }
 
@@ -74,7 +93,7 @@ void iot_log__warn (iot_logger_t * logger, ...)
 {
   va_list args;
   va_start (args, logger);
-  iot_logger_log (logger, IOT_LOG_WARN, args);
+  iot_logger_log ((iot_logger_impl_t*) logger, IOT_LOG_WARN, args);
   va_end (args);
 }
 
@@ -82,7 +101,7 @@ void iot_log__info (iot_logger_t * logger, ...)
 {
   va_list args;
   va_start (args, logger);
-  iot_logger_log (logger, IOT_LOG_INFO, args);
+  iot_logger_log ((iot_logger_impl_t*) logger, IOT_LOG_INFO, args);
   va_end (args);
 }
 
@@ -90,7 +109,7 @@ void iot_log__debug (iot_logger_t * logger, ...)
 {
   va_list args;
   va_start (args, logger);
-  iot_logger_log (logger, IOT_LOG_DEBUG, args);
+  iot_logger_log ((iot_logger_impl_t*) logger, IOT_LOG_DEBUG, args);
   va_end (args);
 }
 
@@ -98,7 +117,7 @@ void iot_log__trace (iot_logger_t * logger, ...)
 {
   va_list args;
   va_start (args, logger);
-  iot_logger_log (logger, IOT_LOG_TRACE, args);
+  iot_logger_log ((iot_logger_impl_t*) logger, IOT_LOG_TRACE, args);
   va_end (args);
 }
 
@@ -108,31 +127,71 @@ void iot_logger_set_level (iot_logger_t * logger, iot_loglevel_t level)
   logger->level = level;
 }
 
-iot_logger_t * iot_logger_alloc (const char * name, iot_loglevel_t level, bool self_start)
+static inline iot_logger_type_t iot_logger_type (iot_log_function_t fn)
 {
-  return iot_logger_alloc_custom (name, level, NULL, iot_log_console, NULL, self_start);
+  if (fn == iot_log_console) return IOT_LOGGER_CONSOLE;
+#if defined (IOT_HAS_FILE) && !defined (_AZURESPHERE_)
+  if (fn == iot_log_file) return IOT_LOGGER_FILE;
+#endif
+  if (fn == iot_log_udp) return IOT_LOGGER_UDP;
+  return IOT_LOGGER_CUSTOM;
 }
 
-iot_logger_t * iot_logger_alloc_custom (const char * name, iot_loglevel_t level, const char * to, iot_log_function_t impl, iot_logger_t * next, bool self_start)
+iot_logger_t * iot_logger_alloc_custom (const char * name, iot_loglevel_t level, const char * to, iot_log_function_t impl, iot_logger_t * next, bool start)
 {
+  static const int yes = 1;
   assert (name && impl);
-  iot_logger_t * logger = calloc (1, sizeof (*logger));
+  iot_logger_type_t type = iot_logger_type (impl);
+  iot_logger_impl_t * logger = calloc (1, sizeof (*logger));
   logger->impl = impl;
+  logger->sock = -1;
   logger->name = strdup (name);
-  logger->to = to ? strdup (to) : NULL;
   logger->save = level;
-  logger->next = next;
-  iot_component_init (&logger->component, IOT_LOGGER_FACTORY, (iot_component_start_fn_t) iot_logger_start, (iot_component_stop_fn_t) iot_logger_stop);
-  if (self_start) iot_logger_start (logger);
-  return logger;
+  logger->next = (iot_logger_impl_t*) next;
+  if (type == IOT_LOGGER_UDP)
+  {
+    const char * sep = strchr (to, ':');
+    if (sep)
+    {
+      char target[17] = { 0 };
+      strncpy (target, to, (size_t) (sep - to));
+      inet_aton (target, &logger->addr.sin_addr);
+      to = sep + 1;
+    }
+    else
+    {
+      logger->addr.sin_addr.s_addr = htonl (INADDR_BROADCAST);
+    }
+    logger->addr.sin_port = htons ((uint16_t) atoi (to));
+    logger->sock = socket (AF_INET, SOCK_DGRAM, 0);
+    if (sep == NULL) setsockopt (logger->sock, SOL_SOCKET, SO_BROADCAST, (char *) &yes, sizeof (yes));
+  }
+#if defined (IOT_HAS_FILE) && !defined (_AZURESPHERE_)
+  else if (type == IOT_LOGGER_FILE)
+  {
+    logger->fd = fopen (to, "a");
+  }
+#endif
+  iot_component_init (&logger->base.component, IOT_LOGGER_FACTORY, (iot_component_start_fn_t) iot_logger_start, (iot_component_stop_fn_t) iot_logger_stop);
+  if (start) iot_logger_start (&logger->base);
+  return &logger->base;
+}
+
+iot_logger_t * iot_logger_alloc (const char * name, iot_loglevel_t level, bool start)
+{
+  return iot_logger_alloc_custom (name, level,NULL, iot_log_console, NULL, start);
 }
 
 void iot_logger_free (iot_logger_t * logger)
 {
-  if (logger && (logger != &iot_logger_dfl) && iot_component_dec_ref (&logger->component))
+  iot_logger_impl_t * impl = (iot_logger_impl_t*) logger;
+  if (impl && (impl != &iot_logger_dfl) && iot_component_dec_ref (&logger->component))
   {
-    free (logger->name);
-    free (logger->to);
+    free (impl->name);
+    close (impl->sock);
+#if defined (IOT_HAS_FILE) && !defined (_AZURESPHERE_)
+    if (impl->fd) fclose (impl->fd);
+#endif
     iot_component_fini (&logger->component);
     free (logger);
   }
@@ -146,54 +205,68 @@ void iot_logger_add_ref (iot_logger_t * logger)
 void iot_logger_start (iot_logger_t * logger)
 {
   assert (logger);
+  iot_logger_impl_t * impl = (iot_logger_impl_t*) logger;
   iot_component_set_running (&logger->component);
-  logger->level = logger->save;
+  logger->level = impl->save;
 }
 
 void iot_logger_stop (iot_logger_t * logger)
 {
   assert (logger);
+  iot_logger_impl_t * impl = (iot_logger_impl_t*) logger;
   iot_component_set_stopped (&logger->component);
-  logger->save = logger->level;
+  impl->save = logger->level;
   logger->level = IOT_LOG_NONE;
 }
 
-iot_logger_t * iot_logger_next (iot_logger_t * logger)
-{
-  assert (logger);
-  return logger->next;
-}
-
-static inline void iot_logger_log_to_fd (iot_logger_t * logger, FILE * fd, iot_loglevel_t level, uint64_t timestamp, const char *message)
+static inline size_t iot_logger_format_log (iot_logger_impl_t * logger, iot_loglevel_t level, uint64_t timestamp, const char * message)
 {
   char tname[IOT_PRCTL_NAME_MAX] = { 0 };
 #ifdef IOT_HAS_PRCTL
   prctl (PR_GET_NAME, tname);
 #endif
-  iot_component_lock (&logger->component);
-#ifdef _AZURESPHERE_
-  Log_Debug ("[%s:%" PRIu64 ":%s:%s] %s\n", tname, (uint64_t) timestamp, logger->name ? logger->name : "default", iot_log_levels[level], message);
-#else
-  fprintf (fd, "[%s:%" PRIu64 ":%s:%s] %s\n", tname, timestamp, logger->name ? logger->name : "default", iot_log_levels[level], message);
-#endif
-  iot_component_unlock (&logger->component);
+  return (size_t) snprintf (logger->buff, sizeof (logger->buff), "[%s:%" PRIu64 ":%s:%s] %s\n", tname, timestamp, logger->name, iot_log_levels[level], message);
 }
 
-#if defined(IOT_HAS_FILE) && !defined(_AZURESPHERE_)
-void iot_log_file (iot_logger_t * logger, iot_loglevel_t level, uint64_t timestamp, const char * message)
+static inline void iot_logger_log_to_fd (iot_logger_impl_t * logger, FILE * fd, iot_loglevel_t level, uint64_t timestamp, const char *message)
 {
-  FILE * fd = fopen (logger->to, "a");
-  if (fd)
+  iot_component_lock (&logger->base.component);
+  if (iot_logger_format_log (logger, level, timestamp, message))
   {
-    iot_logger_log_to_fd (logger, fd, level, timestamp, message);
-    fclose (fd);
+#ifdef _AZURESPHERE_
+    Log_Debug ("%s", logger->buff);
+#else
+    fprintf (fd, "%s", logger->buff);
+#endif
   }
+  iot_component_unlock (&logger->base.component);
+}
+
+#if defined (IOT_HAS_FILE) && !defined (_AZURESPHERE_)
+extern void iot_log_file (iot_logger_t * logger, iot_loglevel_t level, uint64_t timestamp, const char * message)
+{
+  iot_logger_impl_t * impl = (iot_logger_impl_t*) logger;
+  if (impl->fd) iot_logger_log_to_fd (impl, impl->fd, level, timestamp, message);
 }
 #endif
 
 extern void iot_log_console (iot_logger_t * logger, iot_loglevel_t level, uint64_t timestamp, const char * message)
 {
-  iot_logger_log_to_fd (logger, (level > IOT_LOG_WARN) ? stdout : stderr, level, timestamp, message);
+  iot_logger_log_to_fd ((iot_logger_impl_t*) logger, (level > IOT_LOG_WARN) ? stdout : stderr, level, timestamp, message);
+}
+
+/* iot_log_udp: To is either "<host>:<port>" or "<port>". Latter form means broadcast. */
+
+extern void iot_log_udp (iot_logger_t * logger, iot_loglevel_t level, uint64_t timestamp, const char * message)
+{
+  iot_logger_impl_t * impl = (iot_logger_impl_t*) logger;
+  iot_component_lock (&logger->component);
+  if (impl->sock != -1)
+  {
+    size_t len = iot_logger_format_log (impl, level, timestamp, message);
+    if (len > 0) sendto (impl->sock, impl->buff, len, 0, (struct sockaddr *) &impl->addr, sizeof (struct sockaddr_in));
+  }
+  iot_component_unlock (&logger->component);
 }
 
 #ifdef IOT_BUILD_COMPONENTS
@@ -219,8 +292,6 @@ static iot_loglevel_t iot_logger_config_level (const iot_data_t * map)
 
 static iot_component_t * iot_logger_config (iot_container_t * cont, const iot_data_t * map)
 {
-  iot_logger_t * next;
-  iot_logger_t * logger;
   iot_log_function_t impl = iot_log_console; /* log to stderr or stdout */
   iot_loglevel_t level = iot_logger_config_level (map);
   const char * to = iot_data_string_map_get_string (map, "To");
@@ -231,11 +302,16 @@ static iot_component_t * iot_logger_config (iot_container_t * cont, const iot_da
     impl = iot_log_file; /* Log to file */
     to += 5;
   }
+  else
 #endif
-  next = (iot_logger_t*) iot_container_find_component (cont, iot_data_string_map_get_string (map, "Next"));
-  bool self_start = iot_data_string_map_get_bool (map, "Start", true);
-  logger = iot_logger_alloc_custom (iot_data_string_map_get_string (map, "Name"), level, to, impl, next, self_start);
-  return &logger->component;
+  if (to && strncmp (to, "udp:", 4) == 0 && strlen (to) > 4)
+  {
+    impl = iot_log_udp; /* Log to udp */
+    to += 4;
+  }
+  iot_logger_t * next = (iot_logger_t*) iot_container_find_component (cont, iot_data_string_map_get_string (map, "Next"));
+  bool start = iot_data_string_map_get_bool (map, "Start", true);
+  return (iot_component_t*) iot_logger_alloc_custom (iot_data_string_map_get_string (map, "Name"), level, to, impl, next, start);
 }
 
 static bool iot_logger_reconfig (iot_component_t * comp, iot_container_t * cont, const iot_data_t * map)